+++ cypress/integration/members/home.spec.js
@@ -1,4 +1,16 @@
+import {
+  CreateGig,
+  DeleteSignup,
+  SignupDetails,
+  AddInstrument,
+  RemoveInstruments,
+  InstrumentsOnGig,
+} from "../../database/gigs";
 import { CreateUser, HASHED_PASSWORDS } from "../../database/users";
+
+const click = ($el) => $el.click(); // For retrying clicks, see https://www.cypress.io/blog/2019/01/22/when-can-the-test-click/
+
+let colors = {};
 
 describe("members' home page", () => {
   before(() => {
@@ -28,5 +40,209 @@ describe("members' home page", () => {
     cy.login("cypress_user", "abc123");
     cy.visit("/members");
     cy.get("h1").contains("Members");
+  });
+});
+
+describe("gig signup", () => {
+  before(() => {
+    cy.visit("/");
+    cy.executeMutation(CreateGig, {
+      variables: {
+        id: 15274,
+        title: "Cypress Demo Gig",
+        type: 1,
+        adminsOnly: false,
+        allowSignups: true,
+        date: "2020-07-17",
+      },
+    });
+    cy.executeMutation(CreateUser, {
+      variables: {
+        id: 27250,
+        username: "cypress_user",
+        saltedPassword: HASHED_PASSWORDS.abc123,
+        admin: 9,
+        email: "cypress.user@cypress.io",
+        firstName: "Cypress",
+        lastName: "User",
+      },
+    });
+    cy.executeMutation(CreateUser, {
+      variables: {
+        id: 32747,
+        username: "cypress",
+        saltedPassword: HASHED_PASSWORDS.abc123,
+        admin: 1,
+        email: "cy@press.io",
+        firstName: "Cypress",
+        lastName: "Webmaster",
+      },
+    });
+    cy.executeMutation(DeleteSignup, {
+      variables: {
+        userId: 27250,
+        gigId: 15274,
+      },
+    });
+    cy.cssProperty("--positive").then((positive_) => (colors.positive = positive_));
+    cy.cssProperty("--negative").then((negative_) => (colors.negative = negative_));
+    cy.cssProperty("--neutral").then((neutral_) => (colors.neutral = neutral_));
+  });
+
+  beforeEach(() => {
+    cy.clock(Cypress.DateTime.fromISO("2020-07-07T02:00").valueOf());
+    cy.login("cypress_user", "abc123");
+  });
+
+  it("allows a user to sign up to a gig", () => {
+    cy.visit("/members");
+    cy.get(`[data-test="gig-15274-signup-yes"]`).should("not.have.color", colors.positive);
+    cy.get(`[data-test="gig-15274-signup-yes"]`).pipe(click).should("have.color", colors.positive);
+
+    cy.executeQuery(SignupDetails, {
+      variables: {
+        userId: 27250,
+        gigId: 15274,
+      },
+    })
+      .its("cucb_gigs_lineups_by_pk")
+      .should("include", {
+        user_available: true,
+        user_only_if_necessary: false,
+      });
+
+    cy.get(`[data-test="gig-15274-signup-maybe"]`).should("not.have.color", colors.netural);
+    cy.get(`[data-test="gig-15274-signup-maybe"]`).click().should("have.color", colors.neutral);
+
+    cy.executeQuery(SignupDetails, {
+      variables: {
+        userId: 27250,
+        gigId: 15274,
+      },
+    })
+      .its("cucb_gigs_lineups_by_pk")
+      .then((res) => {
+        cy.log("Checking user available");
+        expect(res.user_available).to.equal(true);
+        cy.log("Checking user if necessary");
+        expect(res.user_only_if_necessary).to.equal(true);
+      });
+
+    cy.executeQuery(SignupDetails, {
+      variables: {
+        userId: 27250,
+        gigId: 15274,
+      },
+    })
+      .its("cucb_gigs_lineups_by_pk")
+      .should("include", {
+        user_available: true,
+        user_only_if_necessary: true,
+      });
+
+    cy.get(`[data-test="gig-15274-signup-no"]`).should("not.have.color", colors.negative);
+    cy.get(`[data-test="gig-15274-signup-no"]`).click().should("have.color", colors.negative);
+
+    cy.executeQuery(SignupDetails, {
+      variables: {
+        userId: 27250,
+        gigId: 15274,
+      },
+    })
+      .its("cucb_gigs_lineups_by_pk")
+      .should("include", {
+        user_available: false,
+      });
+  });
+
+  context("without_instruments", () => {
+    before(() => {
+      cy.executeMutation(RemoveInstruments, {
+        variables: {
+          userId: 27250,
+        },
+      });
+      cy.login("cypress_user", "abc123");
+      cy.visit("/members");
+    });
+
+    it("links to user profile when editing instruments", () => {
+      cy.intercept("POST", "/v1/graphql", {
+        fixture: "gig/signup/yes.json",
+      }).as("signup");
+
+      cy.get(`[data-test="gig-15274-signup-yes"]`).pipe(click).should("have.color", colors.positive);
+
+      cy.wait("@signup");
+
+      cy.get(`[data-test="gig-15274-signup-edit"]`).click();
+
+      cy.get(`a[href="/members/user"]`).should("exist");
+    });
+  });
+
+  context("with instruments", () => {
+    before(() => {
+      cy.executeMutation(AddInstrument, {
+        variables: {
+          userId: 27250,
+          instrumentId: 53,
+        },
+      });
+      cy.executeMutation(AddInstrument, {
+        variables: {
+          userId: 27250,
+          instrumentId: 20,
+        },
+      });
+      cy.login("cypress_user", "abc123");
+      cy.visit("/members");
+    });
+
+    it("allows adding and editing of instruments", () => {
+      cy.intercept({
+        method: "POST",
+        url: "/v1/graphql",
+      }).as("graphqlRequest");
+
+      cy.executeQuery(InstrumentsOnGig, {
+        variables: {
+          userId: 27250,
+          gigId: 15274,
+        },
+      })
+        .its("cucb_gigs_lineups_instruments_aggregate.aggregate.count")
+        .should("equal", 0);
+
+      cy.get(`[data-test="gig-15274-signup-yes"]`).pipe(click).should("have.color", colors.positive);
+
+      cy.wait("@graphqlRequest");
+
+      cy.get(`[data-test="gig-15274-signup-edit"]`).click();
+
+      cy.get(`[data-test="gig-15274-signup-instrument-53-toggle"]`).click();
+
+      cy.get(`[data-test="gig-15274-signup-save"]`).click();
+
+      cy.wait("@graphqlRequest");
+
+      cy.get(`[data-test="gig-15274-signup-instruments-selected"]`).contains("Trombone").should("exist");
+
+      cy.get(`[data-test="gig-15274-signup-instruments-selected"]`).contains("Wind Synth").should("not.exist");
+
+      cy.get(`[data-test="gig-15274-signup-edit"]`).click();
+
+      cy.get(`[data-test="gig-15274-signup-instrument-20-toggle"]`).click();
+
+      cy.get(`[data-test="gig-15274-signup-instrument-53-toggle"]`).click();
+
+      cy.get(`[data-test="gig-15274-signup-save"]`).click();
+
+      cy.wait("@graphqlRequest");
+
+      cy.get(`[data-test="gig-15274-signup-instruments-selected"]`).contains("Trombone").should("not.exist");
+
+      cy.get(`[data-test="gig-15274-signup-instruments-selected"]`).contains("Wind Synth").should("exist");
+    });
   });
 });
